Completely refactor the code base to be cleaner
