+++ packages/app/src/App.tsx
@@ -36,7 +36,7 @@ import { AppRouter, FlatRoutes } from '@backstage/core-app-api';
 import { CatalogGraphPage } from '@backstage/plugin-catalog-graph';
 import { RequirePermission } from '@backstage/plugin-permission-react';
 import { catalogEntityCreatePermission } from '@backstage/plugin-catalog-common/alpha';
-import { AISearchFrontendPage } from '@redhatinsights/backstage-plugin-ai-search-frontend';
+import { ConvoFrontendPage } from '@redhatinsights/backstage-plugin-ai-search-frontend';
 
 const app = createApp({
   apis,
@@ -96,7 +96,7 @@ const routes = (
     </Route>
     <Route path="/settings" element={<UserSettingsPage />} />
     <Route path="/catalog-graph" element={<CatalogGraphPage />} />
-    <Route path="/ai-search-frontend" element={<AISearchFrontendPage />} />
+    <Route path="/convo" element={<ConvoFrontendPage />} />
   </FlatRoutes>
 );
