+++ packages/app/src/components/Root/Root.tsx
@@ -72,7 +72,7 @@ export const Root = ({ children }: PropsWithChildren<{}>) => (
           pluralTitle="My Groups"
           icon={GroupIcon}
         />
-        <SidebarItem icon={ExtensionIcon} to="ai-search-frontend" text="AI Search" />
+        <SidebarItem icon={ExtensionIcon} to="convo" text="Convo" />
         <SidebarItem icon={ExtensionIcon} to="api-docs" text="APIs" />
         <SidebarItem icon={LibraryBooks} to="docs" text="Docs" />
         <SidebarItem icon={CreateComponentIcon} to="create" text="Create..." />
