+++ plugins/ai-search-frontend/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@redhatinsights/backstage-plugin-ai-search-frontend",
-  "version": "0.4.2",
+  "name": "@redhatinsights/backstage-plugin-convo-frontend",
+  "version": "1.0.0",
   "main": "src/index.ts",
   "types": "src/index.ts",
   "license": "Apache-2.0",
