--- plugins/ai-search-frontend/src/components/AISearchComponent/index.ts
@@ -1,1 +1,0 @@
-export { AISearchComponent } from './AISearchComponent';
