+++ plugins/ai-search-frontend/src/components/Citations/Citations.tsx
