+++ plugins/ai-search-frontend/src/components/Convo/index.ts
@@ -1,0 +1,1 @@
+export { Convo } from './Convo';
