+++ plugins/ai-search-frontend/src/routes.ts
@@ -1,5 +1,5 @@
 import { createRouteRef } from '@backstage/core-plugin-api';
 
 export const rootRouteRef = createRouteRef({
-  id: 'ai-search-frontend',
+  id: 'convo-frontend',
 });
